requirements.txt
dnspython
pydot
requests
